requirements.txt
python-dotenv==1.1.0
PyYAML==6.0.2
semantic_kernel==1.29.0
typer==0.16.0
uv
azure-identity>=1.23.0
azure-mgmt-resource>=24.0.0
httpx>=0.28.1
mcp[cli]>=1.11.0
pandas>=2.3.1
pyodbc>=5.2.0
sqlalchemy>=2.0.41